+++ resources/views/laravel-examples/user-management.blade.php
@@ -58,7 +58,8 @@
                                     </td>
                                     <td>
                                         <div>
-                                            <img src="../assets/img/team-2.jpg" class="avatar avatar-sm me-3">
+                                            
+                                            <img src="{{ asset('/assets/img/team-2.jpg') }}" class="avatar avatar-sm me-3">
                                         </div>
                                     </td>
                                     <td class="text-center">