+++ resources/views/layouts/user_type/guest.blade.php
@@ -5,7 +5,7 @@
         @include('layouts.navbars.guest.nav')
         @yield('content') 
     @else
-        <div class="container position-sticky z-index-sticky top-0">
+        <div class="container position-sticky z-index-sticky top-0 0104092025010518120114070701230902152315">
             <div class="row">
                 <div class="col-12">
                     @include('layouts.navbars.guest.nav')