+++ CHANGELOG.md
@@ -1,12 +1,12 @@
 # Change Log
-All notable changes to `Soft UI Dashboard Laravel` will be documented in this file.
+All notable changes to `E-Procurement VhiWEB` will be documented in this file.
 
 ## [1.1.0]
 ### Update to Laravel 11.x
 
 ## [1.0.0]
 ### Original Release
-- Soft UI Dashboard Free
+- E-Procurement VhiWEB Free
 - Login
 - Register
 - Forgot password